Task 7: Splitted code
