+++ tdd_intro/homework/07_filecopier/FilesystemUtils.h
@@ -1,0 +1,3 @@
+#include "IFilesystem.h"
+
+std::error_code CopyDirectory(IFilesystem* fs, const PathType& src, const PathType& dst);
