+++ tdd_intro/homework/07_filecopier/IFilesystem.h
@@ -1,0 +1,9 @@
+#include <string>
+
+using PathType = std::string;
+
+class IFilesystem
+{
+public:
+    virtual bool Exists(const PathType& path) const = 0;
+};
